requirements.txt
python-dotenv
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-google
livekit-plugins-noise-cancellation
